requirements-dev.txt
# Test + tooling dependencies. Production has none — the skill itself is pure markdown.
pytest>=8.0
pyyaml>=6.0
playwright>=1.45

# Used by the autonomous self-iteration loop
anthropic>=0.40
